Arranja bug na função externa pow. Adiciona pasta com programas escritos em plush

+++ c_functions.c
@@ -103,5 +103,5 @@ int** int_array_array(int size) {
 }
 
 int pow_int(int b, int e) {
-  return pow(b, e);
+  return (int) pow(b, e);
 }
